✨ feat: add file upload size, refactor execution logic with improved error handling and HTTP status

+++ apps/server/src/services/dockerService.ts
@@ -2,7 +2,26 @@ import { exec } from 'child_process';
 import { tempDir } from '../utils/fileSystem';
 import { logger } from '../utils/logger';
 
-export const runDockerContainer = (filePath: string, fileName: string): Promise<string> => {
+const checkDockerRunning = (): Promise<boolean> => {
+  return new Promise((resolve) => {
+    exec('docker info', (error) => {
+      if (error) {
+        logger.error('Docker is not running');
+        resolve(false);
+      } else {
+        logger.info('Docker is running');
+        resolve(true);
+      }
+    });
+  });
+};
+
+export const runDockerContainer = async (filePath: string, fileName: string): Promise<string> => {
+  const isDockerRunning = await checkDockerRunning();
+  if (!isDockerRunning)
+    throw new Error(`Docker is not running. Please start Docker and try again.`);
+  logger.info('Docker is running, proceeding with container execution');
+
   const containerName = `python-script-runner-${Date.now()}`;
   const command = `docker run --name ${containerName} --rm -v ${tempDir}:/code python:3.9-slim python /code/${fileName}`;
   const timeout = 60 * 1000; // 60 seconds
